feat: swagger를 위한 dto 생성
검색 조건을 받는 dto
검색 결과를 전달하는 dto
상세 조회 결과를 전달하는 dto
